Set initial test time properly
